+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Next version
 - Added a warning when registering a proxy plugin with the ``RegionRenderer``
   with ``fetch=True``. Users should only register the concrete plugin directly
   for fetching, otherwise duplicated plugins are very likely while rendering.
+- Added ``usedforsecurity=False`` to our usage of ``hashlib.md5``.
 
 
 5.4 (2025-06-19)